Make --resident take a double argument, and fix memory deallocation and locking bugs.
darcs-hash:20071213080632-d1b56-1f631a686e8e9564ebc8b1b088d6897d7500a9d3.gz

+++ d2/image.cc
@@ -24,4 +24,4 @@
  * See image.h for details on these variables.
  */
 
-unsigned int image::resident = 0;
+double image::resident = 0;
